Change MCD conversion to FLOAT16 for ANE compatibility
Changed compute_precision from FLOAT32 to FLOAT16 in MultiCodeDecoder
conversion. This will enable ANE execution for 546 ops currently stuck
on CPU due to 'Invalid output tensor format: fp32'.

Expected impact: 0% → ~93% ANE usage (matching CodeDecoder), 9x speedup.

Note: Previous FLOAT32 was to avoid NaN from large CodeDecoder outputs
(~100 range), but profiling shows this blocks ALL ANE usage. If NaN
occurs with FP16, we'll need to clip/normalize CodeDecoder outputs in
Swift before feeding to MCD.

Requires reconversion: cd models/tts/qwen3/coreml && uv run python \
  convert/convert_argmax_multi_code_decoder.py --quantize-w8

+++ models/tts/qwen3/coreml/convert/convert_argmax_multi_code_decoder.py
@@ -357,17 +357,17 @@ def main():
         ct.TensorType(name="new_value_cache", dtype=np.float16),
     ]
 
-    # FLOAT32 precision is required here because the CodeDecoder outputs
-    # post-RMSNorm hidden_states with large outlier values (up to ~100 due to
-    # RMSNorm weights up to ~20). In FLOAT16 precision, the Q/K/V matmuls inside
-    # the MCD layers produce NaN from intermediate overflow with these inputs.
-    # The MCD is only 5 layers so the perf impact of fp32 is minimal.
+    # PERFORMANCE FIX: Changed from FLOAT32 to FLOAT16 for ANE compatibility.
+    # Previous comment claimed fp16 caused NaN due to CodeDecoder outputting
+    # large values (~100), but profiling shows FLOAT32 blocks ALL ANE usage (0%).
+    # CodeDecoder achieves 93% ANE with fp16, so we test if MCD can too.
+    # If NaN occurs, we'll need to clip/normalize CodeDecoder outputs in Swift.
     ml_model = ct.convert(
         traced,
         inputs=inputs,
         outputs=outputs,
         minimum_deployment_target=ct.target.iOS18,
-        compute_precision=ct.precision.FLOAT32,
+        compute_precision=ct.precision.FLOAT16,
     )
 
     # 6. W8A16 quantization
